pyproject.toml
[project]
name = "kbju-bot-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "pyzbar>=0.1.9",
]
